requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlmodel==0.0.16
python-dotenv==1.0.0
bcrypt==4.1.1
jinja2==3.1.2
requests==2.31.0